New: 增加MinioURL注解，可以返回有效时长链接

+++ ingot-framework/ingot-minio-common/src/main/java/com/ingot/framework/minio/common/MinioItem.java
@@ -1,11 +1,11 @@
 package com.ingot.framework.minio.common;
 
+import java.time.ZonedDateTime;
+
 import io.minio.messages.Item;
 import io.minio.messages.Owner;
 import lombok.AllArgsConstructor;
 import lombok.Data;
-
-import java.time.ZonedDateTime;
 
 /**
  * <p>Description  : MinioItem.桶中的对象信息</p>
@@ -28,7 +28,7 @@ public class MinioItem {
         this.objectName = item.objectName();
         this.lastModified = item.lastModified();
         this.etag = item.etag();
-        this.size = (long) item.size();
+        this.size = item.size();
         this.storageClass = item.storageClass();
         this.owner = item.owner();
         this.type = item.isDir() ? "directory" : "file";
